requirements.txt
Django~=3.1.5
djangorestframework~=3.12.2
djoser~=2.1.0
drf-nested-routers~=0.92.5
